combine DatabaseSetup and InputValidator classes
